fix(community): widen search document versions

+++ apps/api/src/password_detective/db/models/community.py
@@ -919,7 +919,7 @@ class CommunitySearchDocument(Base):
         Enum(CommunitySearchSource, native_enum=False, length=16), index=True
     )
     source_id: Mapped[str] = mapped_column(String(36), index=True)
-    document_version: Mapped[int] = mapped_column(Integer, default=1)
+    document_version: Mapped[int] = mapped_column(BigInteger, default=1)
     title: Mapped[str] = mapped_column(String(120))
     body: Mapped[str] = mapped_column(Text, default="")
     username: Mapped[str | None] = mapped_column(String(64), nullable=True, index=True)
@@ -946,7 +946,7 @@ class CommunitySearchOutbox(Base):
         Enum(CommunitySearchSource, native_enum=False, length=16), index=True
     )
     source_id: Mapped[str] = mapped_column(String(36), index=True)
-    document_version: Mapped[int] = mapped_column(Integer, default=1)
+    document_version: Mapped[int] = mapped_column(BigInteger, default=1)
     dedupe_key: Mapped[str] = mapped_column(String(160), unique=True)
     status: Mapped[CommunitySearchOutboxStatus] = mapped_column(
         Enum(CommunitySearchOutboxStatus, native_enum=False, length=16),
